third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-dom-z-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'z' attribute of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.getAttribute('z') is "0"
PASS spotLightElement.getAttribute('z') is "100"
PASS successfullyParsed is true

TEST COMPLETE
